feat(gcov coverage): added a check to handel gcov8 output
Gcov 8 changes the way the outputs work, this requires a differnt way of gerring the files. As the
versions prior to 8 need to be processed in the older manner a version check has been added to
select between these to options.

feat(coverage): changed the gcov command

Change the GCOV command to the command used in the other parts of the script

feat(coverage): moved the GCOV call and updated the description

The GCOV call was moved to in the version check to allow gcov 8.0.0 to generate branch information.
The description for this gcov call in 8.0.0 or higher is updated to comply with the executed
command.

feat(coverage): change version check

feat(coverage): updated the description

feat(coverage): added a check for the 9.1.0 version

in the 9.1.0 version the file name changes form <name>.gcov to <name>.gcov.json.gz